2.2.Hooking into Symfony with an Event Subscriber
This would work, but starting in Symfony 4.3, instead of using this made-up kernel.request string, you can pass the event class name, which in this case is RequestEvent::class.
More and more, you'll see documentation that tells you to listen to an event class like this, instead of a random string.
Now, create the function: public function onKernelRequest(). Inside, dump and die it's alive!.
Cool! With any luck, Symfony will call our event listener very early on and it will kill the page. Close the profiler, refresh and... it's alive! Well actually, it's dead, but ya know... that's what we wanted!