2.3.Hooking into Symfony with an Event Subscriber{Logging in the Listener and Controller}
Logging in the Listener and Controller
To make the class more interesting, let's log something! You know the drill: add public function __construct() with LoggerInterface $logger. I'll hit Alt+Enter and go to initialize fields as a lazy way to create the property and set it down here.

In the method, add $this->logger->info() with:

  I'm logging SUPER early on the request!

To compare this to logging in a controller, go back to ArticleController. On the homepage action, autowire a  argument and say $logger->info():

  Inside the controller!

We expect that the listener will be called first because the RequestEvent, also known as kernel.request, happens before the controller is executed. Refresh the page. It works... and once again, open the profiler in a new tab, click Logs and... perfect! First our listener log and then the controller.
And you can now see our subscriber inside the performance section! Make sure you have the threshold set to 0. Let's see... there it is: UserAgentSubscriber. And then down... way after that... is the controller.