7.2.Who Creates the Controller & Gives it the Container? {The Final Callable Controller Result; Controllers: Boring Services}
The Final Callable Controller Result
Scroll back up in ControllerResolver to getController(). This is all a long way of saying that our controller string - this App\Controller\ArticleController::homepage - is split into two pieces, the service is fetched from the container, and it's returned from here in a callable format.
Close both of the controller resolver classes and head back to HttpKernel. Let's see what this final $controller looks like. After the if, dd($controller)
Ok, move over... and refresh. That's it! The weird PHP callable syntax: an array where the 0 index is an ArticleController object, and the 1 index is the string homepage.

Controllers: Boring Services
Go ahead and remove that dd(). So... this is beautiful. Our controller is a boring service object: there's nothing special about it at all. Need to use a service like the logger? No problem! In ArticleController, add another argument to the constructor: LoggerInterface $logger. I'll hit Alt + Enter and go to "Initialize Fields" to create that property and set it. To prove it's working, let's say $this->logger->info('Controller instantiated! ')
Move over, refresh, click a link to open the profiler and go to the Logs section. Cool. The first log is from our listener to kernel.request, then our controller is instantiated and then it's executed.
So yea! Our controller is a boring service. Well, it does have that superpower where you can autowire services into controller methods - but we'll learn how that works in a few minutes.
I do have one more question, though. The controller is full of shortcut methods like $this->render(). How does that work? We never injected the twig service... so how is our "boring, normal service" using something that we didn't inject? How is it getting the twig service?